requirements.txt
allure-behave==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
behave==1.2.6
browserstack-local==1.2.3
parse==1.19.0
parse-type==0.5.2
Paver==1.3.4
pluggy==1.0.0
psutil==5.8.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
